Deprecate #local_deps_path
The config method #local_deps_path is a convenient and short way to
share a folder of Babushka deps with the virtual machine. However,
it doesn't really do much (it's basically just a convenience method),
and it's difficult to customise (it doesn't support NFS, sharing
multiple directories, or changing the ID of the shared folder).
Somebody unfamiliar with this plugin who reads this in use would have
to check the documentation or source code for this plugin, to see that
it's just sharing a folder with the virtual machine.

This change outputs a warning to Vagrant's log if #local_deps_path is
used, explaining that the method is deprecated, and offering an
equivalent configuration line that can be used to have the exact same
effect as the old method did. It also allows the method to continue
working exactly as it has. The purpose is to notify anybody using this
method that it's deprecated, and explain the benefits of sharing the
folder manually.